requirements.txt
pandas
python-dotenv
groq
sentence-transformers
scikit-learn
joblib